pyproject.toml
[project]
name = "aoe4-discord"
version = "0.1.0"
readme = "README.md"
description = "AOE4 Discord Bot"
authors = [
  {name = "Ryan de Marigny", email = "ryandemarigny@gmail.com"},
]
license = {file = "LICENSE"}
keywords = ["aoe4", "aoe4-bot", "discord", "age of empires", "discord bot"]

classifiers = [
  "Development Status :: 3 - Alpha",

  "Intended Audience :: Developers",

  "License :: OSI Approved :: MIT License",

  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "aiohttp >= 3.9.0",
  "discord.py >= 2.3.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
